Extract AdyenTextField decoration box into a separate composable
COSDK-506
